requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==4.2.23
django-jazzmin==3.0.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
whitenoise==6.9.0
